feat: move shared unit test codes to its own file
